% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_response_variable}
\alias{convert_response_variable}
\title{Convert response variable to appropriate format for neural networks}
\usage{
convert_response_variable(y)
}
\arguments{
\item{y}{Response variable from model.frame}
}
\value{
Numeric vector suitable for neural network training
}
\description{
Convert response variable to appropriate format for neural networks
}
