% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_methods.R
\name{generate_predictions.keras.engine.training.Model}
\alias{generate_predictions.keras.engine.training.Model}
\title{Generate Predictions for Keras Models}
\usage{
\method{generate_predictions}{keras.engine.training.Model}(model, re, pred.values, pred.type, report.se)
}
\arguments{
\item{model}{A fitted Keras model object}

\item{re}{Should random effects be predicted? (Not applicable for neural networks, ignored)}

\item{pred.values}{Data frame containing predictor values for prediction}

\item{pred.type}{Type of predictions (ignored for neural networks)}

\item{report.se}{Should standard errors be reported? (Not supported for neural networks)}
}
\value{
A data frame with columns "prediction" and "model"
}
\description{
S3 method for generating predictions from Keras models for use with flexplot's
compare.fits function.
}
