% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_response_var}
\alias{set_response_var}
\title{Set Response Variable for Keras Model}
\usage{
set_response_var(model, response_var)
}
\arguments{
\item{model}{Keras model object}

\item{response_var}{Name of the response variable}
}
\value{
The model object with response variable attribute set
}
\description{
Helper function to store the response variable name as a model attribute.
This is useful for the get_terms method.
}
